refactor(bridges): split concrete bridges out of base.py
base.py held the abstract Bridge plus two concrete subclasses. Mirror the
outputs/ convention (base.py = the ABC; one module per concrete):

- WebCamBridge -> webcam.py, reunited with the `python -m ...webcam` subprocess
  it launches (the two halves of the webcam feature now live together)
- CustomBridge -> custom.py
- base.py keeps only Bridge (and drops the now-unused `import sys`)

Public API unchanged: still re-exported from myogestic.bridges.

Verified: ruff + ty + pyright clean, 176 tests, docs 0 warnings.
